+++ activesupport/test/core_ext/object/self_test.rb
@@ -1,0 +1,9 @@
+require 'abstract_unit'
+require 'active_support/core_ext/object'
+
+class Object::SelfTest < ActiveSupport::TestCase
+  test 'self returns self' do
+    object = 'fun'
+    assert_equal object, object.self
+  end
+end