+++ activesupport/CHANGELOG.md
@@ -1,3 +1,9 @@
+*   Added Object#self which returns the object itself. Useful when dealing with a chaining scenario, like Active Record scopes:
+
+        Event.public_send(state.presence_in?([ :trashed, :drafted ]) ? :self).order(:created_at)
+    
+    *DHH*
+
 *   `Object#with_options` executes block in merging option context when
     explicit receiver in not passed.